More project reorganization Allow to the server to use game screen and input system
