+++ BlazorServer/AddonConfigurator.cs
@@ -9,6 +9,7 @@ using System;
 using System.Collections.Generic;
 using System.Text.RegularExpressions;
 using SharedLib;
+using Game;
 
 namespace BlazorServer
 {
