+++ BlazorServer/Pages/FrameConfiguration.razor
@@ -9,6 +9,7 @@
 @using Microsoft.AspNetCore.Html
 @using System.Threading;
 @using SharedLib;
+@using Game;
 
 <style>
     h5 {
@@ -256,7 +257,7 @@ else
 
                                 if (dataFrames.Count == dataFrameMeta.frames)
                                 {
-                                    if(addonReader != null)
+                                    if (addonReader != null)
                                     {
                                         addonReader.Dispose();
                                         addonDataProvider = null;
