+++ BlazorServer/Pages/MiniMapComponent.razor
@@ -1,5 +1,6 @@
 ﻿@using Core
 @using SharedLib;
+@using Game;
 @inject Core.IBotController botController
 @inject Core.IAddonReader addonReader
