+++ BlazorServer/Pages/ScreenshotComponent.razor
@@ -3,6 +3,7 @@
 <img src="data:image/png;base64, @image" alt="Red dot" />
 
 @using SharedLib;
+@using Game;
 @code {
 
     [Parameter]
