+++ Core/Addon/AddonDataProvider/AddonDataProvider.cs
@@ -1,9 +1,10 @@
-﻿using SharedLib;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
 using System.Text;
+using SharedLib;
+using Game;
 
 namespace Core
 {
