+++ Core/BotController.cs
@@ -9,9 +9,10 @@ using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Core.Utils;
 using Core.Database;
 using SharedLib;
+using Game;
+using WinAPI;
 
 namespace Core
 {
