+++ Core/ConfigBotController.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading;
 using SharedLib;
+using Game;
 
 namespace Core
 {
