+++ Core/Core.csproj
@@ -23,7 +23,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="GregsStack.InputSimulatorStandard" Version="1.3.1" />
     <PackageReference Include="Microsoft.AspNetCore.Html.Abstractions" Version="2.2.0" />
     <PackageReference Include="Microsoft.CodeAnalysis.FxCopAnalyzers" Version="2.9.6">
       <PrivateAssets>all</PrivateAssets>
@@ -37,8 +36,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\DataConfig\DataConfig.csproj" />
-    <ProjectReference Include="..\SharedLib\SharedLib.csproj" />
-    <ProjectReference Include="..\WinAPI\WinAPI.csproj" />
+    <ProjectReference Include="..\Game\Game.csproj" />
     <ProjectReference Include="..\WowheadDB\WowheadDB.csproj" />
   </ItemGroup>
