+++ Core/Goals/StopMoving.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Extensions.Logging;
 using System;
 using System.Threading.Tasks;
+using Game;
 
 namespace Core.Goals
 {
