+++ Core/IBotController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Threading;
 using SharedLib;
+using Game;
 
 namespace Core
 {
