+++ Core/MiniMapNodeAlert/MinimapNodeFinder.cs
@@ -4,6 +4,7 @@ using System.Diagnostics;
 using System.Drawing;
 using System.Linq;
 using SharedLib;
+using Game;
 
 #nullable enable
 namespace Core
