+++ Core/NpcFinder/NpcNameFinder.cs
@@ -8,6 +8,7 @@ using System.Drawing;
 using System.Linq;
 using System.Threading.Tasks;
 using SharedLib;
+using Game;
 
 namespace Core
 {
