+++ Game/Input/IMouseInput.cs
@@ -1,7 +1,7 @@
 ﻿using System.Drawing;
 using System.Threading.Tasks;
 
-namespace Core
+namespace Game
 {
     public interface IMouseInput
     {
