+++ Game/Input/InputSimulator.cs
@@ -7,7 +7,7 @@ using GregsStack.InputSimulatorStandard;
 using GregsStack.InputSimulatorStandard.Native;
 using WinAPI;
 
-namespace Core
+namespace Game
 {
     public class InputSimulator : IInput
     {
