+++ Game/Input/InputWindowsNative.cs
@@ -7,7 +7,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using WinAPI;
 
-namespace Core
+namespace Game
 {
     public class InputWindowsNative : IInput
     {
