+++ Game/Input/WowProcessInput.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 using SharedLib;
 using WinAPI;
 
-namespace Core
+namespace Game
 {
     public class WowProcessInput : IMouseInput
     {
