+++ Game/WoWScreen/IWowScreen.cs
@@ -1,4 +1,6 @@
-﻿namespace SharedLib
+﻿using SharedLib;
+
+namespace Game
 {
     public interface IWowScreen : IColorReader, IRectProvider
     {
