+++ MasterOfPuppets.sln
@@ -25,7 +25,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AddonConfig", "AddonConfig\
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WinAPI", "WinAPI\WinAPI.csproj", "{DC3149ED-D37A-4020-B4A0-3AD3B86754F3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Server", "Server\Server.csproj", "{BFB5D6D4-C6FF-4A19-97F7-7C8205AD1FD4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Server", "Server\Server.csproj", "{BFB5D6D4-C6FF-4A19-97F7-7C8205AD1FD4}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Game", "Game\Game.csproj", "{2DD51B2A-AA1A-493C-9370-72B52B74E350}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -73,6 +75,10 @@ Global
 		{BFB5D6D4-C6FF-4A19-97F7-7C8205AD1FD4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BFB5D6D4-C6FF-4A19-97F7-7C8205AD1FD4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BFB5D6D4-C6FF-4A19-97F7-7C8205AD1FD4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2DD51B2A-AA1A-493C-9370-72B52B74E350}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2DD51B2A-AA1A-493C-9370-72B52B74E350}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2DD51B2A-AA1A-493C-9370-72B52B74E350}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2DD51B2A-AA1A-493C-9370-72B52B74E350}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
