+++ Server/DataProvider/DataProvider.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading;
 using Microsoft.Extensions.Logging;
 using SharedLib;
+using Game;
 
 namespace Server
 {
