+++ SharedLib/DataFrameConfig/DataFrameConfiguration.cs
@@ -1,9 +1,10 @@
 ﻿using Newtonsoft.Json;
-using SharedLib;
 using System.Collections.Generic;
 using System.Drawing;
 using System.IO;
 using System.Linq;
+
+#nullable enable
 
 namespace SharedLib
 {
