+++ SharedLib/SharedLib.csproj
@@ -11,7 +11,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\DataConfig\DataConfig.csproj" />
-    <ProjectReference Include="..\WinAPI\WinAPI.csproj" />
   </ItemGroup>
 
 </Project>
